+++ src/main/java/org/eclipse/uprotocol/transport/datamodel/UListener.java
@@ -1,4 +1,4 @@
-package org.eclipse.uprotocol.utransport.datamodel;
+package org.eclipse.uprotocol.transport.datamodel;
 
 import org.eclipse.uprotocol.uri.datamodel.UUri;
