+++ src/main/java/org/eclipse/uprotocol/transport/datamodel/UPriority.java
@@ -1,4 +1,4 @@
-package org.eclipse.uprotocol.utransport.datamodel;
+package org.eclipse.uprotocol.transport.datamodel;
 
 import java.util.Arrays;
 import java.util.Optional;
