+++ src/test/java/org/eclipse/uprotocol/transport/datamodel/UAttributeTest.java
@@ -19,9 +19,13 @@
  * under the License.
  */
 
-package org.eclipse.uprotocol.utransport.datamodel;
+package org.eclipse.uprotocol.transport.datamodel;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
+
+import org.eclipse.uprotocol.transport.datamodel.UAttributes;
+import org.eclipse.uprotocol.transport.datamodel.UMessageType;
+import org.eclipse.uprotocol.transport.datamodel.UPriority;
 import org.eclipse.uprotocol.uri.datamodel.UAuthority;
 import org.eclipse.uprotocol.uri.datamodel.UEntity;
 import org.eclipse.uprotocol.uri.datamodel.UResource;
